# examples/todo-api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: Todo app demonstrating OpenAPI-driven development with appbase
  version: 0.1.0

paths:
  /api/todos:
    get:
      operationId: listTodos
      summary: List all todos for the authenticated user
      security:
        - session: []
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createTodo
      summary: Create a new todo
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Created todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/todos/{id}:
    delete:
      operationId: deleteTodo
      summary: Delete a todo by ID
      security:
        - session: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  schemas:
    Todo:
      type: object
      required: [id, userId, title, done, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        done:
          type: boolean
        createdAt:
          type: string
          format: date-time

    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1

    OkResponse:
      type: object
      properties:
        ok:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: app_session